Fix boulder loading glitch and change how boulder from url is loaded
Because the ctm loader is asynchronous, code was continuing before
the loader came back with a boulder. Climbsim would try and add an
undefined to the scene and then the boulder would be added on the next
attempt to load a boulder, causing "boulder lag!"

Also, to make sure things happen in the right order, instead of having
the router set a session variable, triggering boulder load, we let
Climbsim.init() check the URL parameter from the router. This way it
always happens at the right time (just after scene init).