Remove web worker loading hack to get .nxz loading working
Nexus goes through all the scripts looking for itself and then replaces
its filename with the name of the web worker it wants. unsurprisingly,
that wasn't working consistently with meteor. Hard code the path
instead. Might be worth finding a better solution to this at some point,
e.g. if it ever becomes possible to load modules as web workers.