Adding withSingle to CragDisp to allow displaying current climb name
The Vulcan pattern seems to be that terms are passed to withSingle as
props. So I'm taking the crag id from the route, and then I have to
get it into the props for CragDisp, which means I need to wrap it in
another react component. Thus ends the sordid story of how
CragDispAnnoyingWrapper came to be.